Add files via upload
This python program (ESC_all_motors.py) allows one to calibrate the ESCs and also control the motors via Raspberry Pi's PWM.

Note: DON'T CONNECT THE RED WIRE OF THE ESC TO YOUR PI. IT WILL BURN YOUR PI'S PROCESSOR. Just connect the black wires of the ESCs to ur Pi's ground pins, and the yellow ones to the Pi's generic GPIO pins. You may follow the pin numberings in the program or connect them as per your convenience. The pin numberings in the program are broadcom numbering, not the general pin numbering. If you are not aware of what this is take a look - https://raspberrypi.stackexchange.com/questions/12966/what-is-the-difference-between-board-and-bcm-for-gpio-pin-numbering